+++ rnt_rd-react_homework/src/components/regForm/RegForm.tsx
@@ -5,22 +5,16 @@ import { Link } from 'react-router-dom';
 
 import { fetchRegUser, inpItems, regCreateAccount, useReg } from './regSlice';
 import { IFormData } from 'types/TypesBase';
+import ValidateForm from 'utils/validate/validateForm';
 
 import netflixBg from '../../resources/img/netflix_bg.jpg';
 import './regForm.scss';
 
 const RegForm = () => {
+    const {validateUsername, validatePassword, validateEmail} = ValidateForm();
     const inputItems = useSelector(inpItems);
     const createdAcc = useSelector(useReg);
     const dispatch = useDispatch();
-
-    const validateUsername = (username: string | undefined) => {
-        const usernameRegExp = /^[A-Z0-9]+$/;
-        return username && !usernameRegExp.test(username) ? 'Big letters and numbers' : undefined;
-    };
-    const validatePassword = (password: string | undefined) => {
-        return password && (password.length < 8 || new Set(password).size !== password.length) ? 'Unique symbols / more than 8' : undefined;
-    }
 
     const submitReg = (values: IFormData, form) => {
         dispatch(fetchRegUser(values) as any);
@@ -34,14 +28,11 @@ const RegForm = () => {
 
     const validateReg = (values: IFormData) => {
         const errors: Partial<IFormData> = {};
-        const emailRegExp = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+.[a-zA-Z]{2,4}$/;
         
         errors.username = validateUsername(values.username);
         errors.password = validatePassword(values.password);
+        errors.email = validateEmail(values.email);
 
-        if (values.email && !emailRegExp.test(values.email)) {
-            errors.email = 'Invalid email';
-        }
         if (values.password !== values.confirmPass) {
             errors.confirmPass = 'Password mismatch';
         }
