+++ rnt_rd-react_homework/src/utils/router/privateRoute.tsx
@@ -5,11 +5,7 @@ import { useSelector } from "react-redux";
 import { useLog } from "components/loginForm/loginSlice";
 
 const PrivateRoute = () => {
-    // const isAuthenticated = localStorage.getItem('isAuthenticated');
-    // console.log(isAuthenticated);
-
     const signIn = useSelector(useLog);
-    console.log(signIn)
 
     return (
         signIn ? <Outlet/> : <Navigate to="/login"/>
